added index on _message and _host. now also using new keys for syslog messages

+++ dataset/GitHub_Java/Graylog2.graylog2-server/src/main/java/org/graylog2/database/MongoBridge.java
@@ -23,6 +23,7 @@ package org.graylog2.database;
 import com.mongodb.DBCollection;
 import com.mongodb.BasicDBObject;
 import com.mongodb.BasicDBObjectBuilder;
+import com.mongodb.WriteConcern;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
@@ -59,7 +60,7 @@ public class MongoBridge {
         if(MongoConnection.getInstance().getDatabase().getCollectionNames().contains("messages")) {
             coll = MongoConnection.getInstance().getDatabase().getCollection("messages");
         } else {
-            long messagesCollSize = Long.parseLong(Main.masterConfig.getProperty("messages_collection_size").trim());
+            int messagesCollSize = Integer.parseInt(Main.masterConfig.getProperty("messages_collection_size").trim());
             coll = MongoConnection.getInstance().getDatabase().createCollection("messages", BasicDBObjectBuilder.start().add("capped", true).add("size", messagesCollSize).get());
         }
